Added some additional notes
Added some different notes to the Nasa c style cheat and wanted to inform my self what the objectives and outcomes are suppose to be.
More to follow when I get a free time and I may write a different one. But believe this is a tried and true tested style with military standards.